% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_est.R
\name{acc_est}
\alias{acc_est}
\title{Accuracy Estimation}
\usage{
acc_est(score, X, beta, lambda0, w = NULL)
}
\arguments{
\item{score}{risk score}

\item{X}{predictors}

\item{beta}{regression coefficients}

\item{lambda0}{baseline hazards}

\item{w}{weights}
}
\description{
Accuracy Estimation
}
