% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adph.R
\name{adph}
\alias{adph}
\title{Adjusted Discrete Proportional Hazards (ADPH) Model}
\usage{
adph(time, status, pred, sens, spec)
}
\arguments{
\item{time}{a vector of discrete time (e.g. 1, 2, 3, ...)}

\item{status}{a vector of event status (1 = observed, 0 = censored)}

\item{pred}{a vector/matrix of predictors (e.g. biomarkers)}

\item{sens}{sensitivity (a scalar or vector)}

\item{spec}{specificity (a scalar or vector)}
}
\description{
Adjusted Discrete Proportional Hazards (ADPH) Model
}
