% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adph2.R
\name{adph2}
\alias{adph2}
\title{Adjusted Discrete Proportional Hazards (ADPH) Model}
\usage{
adph2(time, status, pred, lambda0 = NULL, lambda0_s = NULL)
}
\arguments{
\item{time}{a vector of discrete time (e.g. 1, 2, 3, ...)}

\item{status}{a vector of event status (1 = observed, 0 = censored)}

\item{pred}{a vector/matrix of predictors (e.g. biomarkers)}

\item{lambda0}{baseline hazards}

\item{lambda0_s}{...}
}
\description{
Adjusted Discrete Proportional Hazards (ADPH) Model
}
