% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adph2.R
\name{adph2_loglik}
\alias{adph2_loglik}
\title{ADPH2 Log-likelihood}
\usage{
adph2_loglik(beta, lambda0, lambda0_s, t0, d0, X)
}
\arguments{
\item{beta}{regression coefficients}

\item{lambda0}{baseline hazards}

\item{lambda0_s}{...}

\item{t0}{observed event time}

\item{d0}{observed event indicator}

\item{X}{baseline covaraites}
}
\description{
ADPH2 Log-likelihood
}
