% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adph.R
\name{adph_loglik}
\alias{adph_loglik}
\title{ADPH Log-likelihood}
\usage{
adph_loglik(lambda0, beta, sens, spec, t0, d0, X)
}
\arguments{
\item{lambda0}{baseline hazards}

\item{beta}{regression coefficients}

\item{sens}{sensitivities}

\item{spec}{specificities}

\item{t0}{observed event time}

\item{d0}{observed event indicator}

\item{X}{baseline covaraites}
}
\description{
ADPH Log-likelihood
}
