% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{AUC calculation}
\usage{
auc(score, X, beta, lambda0, w = NULL)
}
\arguments{
\item{score}{risk score}

\item{X}{predictors}

\item{beta}{regression coefficients}

\item{lambda0}{baseline hazards}

\item{w}{weights}
}
\description{
AUC calculation
}
