% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dph.R
\name{dph}
\alias{dph}
\title{Standard Discrete Proportional Hazards Model}
\usage{
dph(time, status, pred)
}
\arguments{
\item{time}{a vector of discrete time (e.g. 1, 2, 3, ...)}

\item{status}{a vector of event status (1 = observed, 0 = censored)}

\item{pred}{a vector/matrix of predictors (e.g. biomarkers)}
}
\description{
Standard Discrete Proportional Hazards Model
}
