% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated Data Example}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
\item{time}{discrete time}
\item{status}{event status}
\item{pred1}{first predictor}
\item{pred2}{second predictor}
}
}
\usage{
sim_data
}
\description{
Simulated Data Example
}
\keyword{datasets}
