% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adph2.R
\name{adph2}
\alias{adph2}
\title{Adjusted Discrete Proportional Hazards (ADPH) Model}
\usage{
adph2(
  time,
  status,
  pred,
  init_beta = NULL,
  init_lambda0 = NULL,
  init_lambda0_s = NULL,
  n_ptb = NULL,
  seed = 1,
  show_progress = FALSE,
  maxit = 1000,
  ...
)
}
\arguments{
\item{time}{a vector of discrete time (e.g. 1, 2, 3, ...)}

\item{status}{a vector of event status (1 = observed, 0 = censored)}

\item{pred}{a vector/matrix of predictors (e.g. biomarkers)}

\item{init_beta}{initial values for beta}

\item{init_lambda0}{initial values for lambda0}

\item{init_lambda0_s}{initial values for lambda0_s}

\item{n_ptb}{number of perturbations}

\item{seed}{random number generation seed}

\item{show_progress}{TRUE or FALSE (default)}

\item{maxit}{maximum number of iterations}

\item{...}{additional arguments for controlling model fitting (see code{\link[stats]{optim}})}
}
\description{
Adjusted Discrete Proportional Hazards (ADPH) Model
}
